refactor: modularize app logic
