+++ docs/js/app.js
@@ -1,4 +1,4 @@
-import { $, $$, nowHM } from './utils.js';
+import { $, $$ } from './utils.js';
 import { initTabs } from './tabs.js';
 import { initChips, setChipActive, isChipActive, addChipIndicators } from './chips.js';
 import { initAutoActivate } from './autoActivate.js';
@@ -11,17 +11,17 @@ import { initTopbar } from './components/topbar.js';
 import { initCollapsibles } from './sections.js';
 import { connectSocket, initSessions, fetchUsers, initTheme, saveAll, loadAll } from './sessionManager.js';
 import bodyMap from './bodyMap.js';
-import { generateReport, gksSum } from './report.js';
+import { generateReport } from './report.js';
 import { setupHeaderActions } from './headerActions.js';
 import { TEAM_ROLES } from './constants.js';
-import { initCirculationChecks } from './circulation.js';
+import { initCirculation } from './circulation.js';
+import { setupActivationControls, ensureSingleTeam, updateActivationIndicator } from './activation.js';
+import { initGcs } from './gcs.js';
+import { IMG_CT, IMG_XRAY, LABS, BLOOD_GROUPS, FAST_AREAS } from './config.js';
 export { validateVitals };
 
 /* ===== Imaging / Labs / Team ===== */
-const IMG_CT=['Galvos KT','Kaklo KT','Viso kūno KT'];
-const IMG_XRAY=['Krūtinės Ro','Dubens Ro'];
-const LABS=['BKT','Biocheminis tyrimas','Krešumai','Fibrinogenas','ROTEM','Kraujo grupė','Kraujo dujos'];
-const BLOOD_GROUPS=['0-','0+','A-','A+','B-','B+','AB-','AB+'];
+const LS_MECHANISM_KEY='traumos_mechanizmai';
 
 const imgCtWrap=$('#imaging_ct'); IMG_CT.forEach(n=>{const s=document.createElement('span'); s.className='chip'; s.dataset.value=n; s.textContent=n; addChipIndicators(s); imgCtWrap.appendChild(s);});
 const imgXrayWrap=$('#imaging_xray'); IMG_XRAY.forEach(n=>{const s=document.createElement('span'); s.className='chip'; s.dataset.value=n; s.textContent=n; addChipIndicators(s); imgXrayWrap.appendChild(s);});
@@ -31,7 +31,8 @@ const bloodGroupWrap=$('#bloodGroup'); if(bloodGroupWrap){ BLOOD_GROUPS.forEach(
 const bloodUnitsInput=$('#bloodUnits');
 const addBloodOrderBtn=$('#addBloodOrder');
 if(bloodUnitsInput && bloodGroupWrap && addBloodOrderBtn){
-  addBloodOrderBtn.addEventListener('click',()=>{
+  const addOrder=e=>{
+    e.preventDefault();
     const units=bloodUnitsInput.value.trim();
     const groupEl=$$('.chip',bloodGroupWrap).find(c=>isChipActive(c));
     const group=groupEl?.dataset.value||'';
@@ -47,18 +48,12 @@ if(bloodUnitsInput && bloodGroupWrap && addBloodOrderBtn){
     saveAll();
     bloodUnitsInput.value='';
     $$('.chip',bloodGroupWrap).forEach(c=>setChipActive(c,false));
-  });
+  };
+  addBloodOrderBtn.addEventListener('click',addOrder);
+  bloodUnitsInput.addEventListener('keydown',e=>{ if(e.key==='Enter') addOrder(e); });
 }
-const fastAreas=[
-  {name:'Perikardas', marker:'skystis'},
-  {name:'Dešinė pleura', marker:'skystis ar oras'},
-  {name:'Kairė pleura', marker:'skystis ar oras'},
-  {name:'RUQ', marker:'skystis'},
-  {name:'LUQ', marker:'skystis'},
-  {name:'Dubuo', marker:'skystis'}
-];
 const fastWrap=$('#fastGrid');
-fastAreas.forEach(({name,marker})=>{
+FAST_AREAS.forEach(({name,marker})=>{
   const box=document.createElement('div');
   box.innerHTML=`<label>${name} (${marker})</label><div class="row"><label class="pill red"><input type="radio" name="fast_${name}" value="Yra"> Yra</label><label class="pill"><input type="radio" name="fast_${name}" value="Nėra"> Nėra</label></div>`;
   fastWrap.appendChild(box);
@@ -70,51 +65,31 @@ const teamWrap=$('#teamGrid'); TEAM_ROLES.forEach(r=>{
   teamWrap.appendChild(box);
 });
 
-/* ===== Activation indicator ===== */
-function ensureSingleTeam(){
-  const red=$('#chips_red');
-  const yellow=$('#chips_yellow');
-  const redActive=$$('.chip.active', red).length>0;
-  const yellowActive=$$('.chip.active', yellow).length>0;
-  if(redActive && yellowActive){
-    $$('.chip', yellow).forEach(c=>setChipActive(c,false));
-  }
-}
-function updateActivationIndicator(){
-  const dot=$('#activationIndicator');
-  const redActive=$$('.chip.active', $('#chips_red')).length>0;
-  const yellowActive=$$('.chip.active', $('#chips_yellow')).length>0;
-  dot.classList.remove('red','yellow');
-  dot.style.display='none';
-  if(redActive){ dot.classList.add('red'); dot.style.display='inline-block'; }
-  else if(yellowActive){ dot.classList.add('yellow'); dot.style.display='inline-block'; }
-}
-function setupActivationControls(){
-  const redGroup=$('#chips_red');
-  const yellowGroup=$('#chips_yellow');
-  redGroup.addEventListener('click',e=>{
-    const chip=e.target.closest('.chip');
-    if(!chip) return;
-    if(isChipActive(chip)){
-      $$('.chip', yellowGroup).forEach(c=>setChipActive(c,false));
+function initMechanismList(){
+  const list=$('#gmp_mechanism_list');
+  const input=$('#gmp_mechanism');
+  if(!list||!input) return;
+  const existing=new Set(Array.from(list.options).map(o=>o.value));
+  const stored=JSON.parse(localStorage.getItem(LS_MECHANISM_KEY)||'[]');
+  stored.forEach(v=>{
+    if(!existing.has(v)){
+      const opt=document.createElement('option');
+      opt.value=v;
+      list.appendChild(opt);
+      existing.add(v);
     }
-    ensureSingleTeam();
-    updateActivationIndicator();
-    saveAll();
   });
-  yellowGroup.addEventListener('click',e=>{
-    const chip=e.target.closest('.chip');
-    if(!chip) return;
-    if(isChipActive(chip)){
-      $$('.chip', redGroup).forEach(c=>setChipActive(c,false));
-    }
-    ensureSingleTeam();
-    updateActivationIndicator();
-    saveAll();
+  input.addEventListener('change',()=>{
+    const val=input.value.trim();
+    if(!val||existing.has(val)) return;
+    const opt=document.createElement('option');
+    opt.value=val;
+    list.appendChild(opt);
+    existing.add(val);
+    stored.push(val);
+    localStorage.setItem(LS_MECHANISM_KEY, JSON.stringify(stored));
   });
 }
-window.updateActivationIndicator=updateActivationIndicator;
-window.ensureSingleTeam=ensureSingleTeam;
 
 /* ===== Save / Load ===== */
 function expandOutput(){
@@ -132,81 +107,8 @@ function debounce(fn, delay){
   };
 }
 const saveAllDebounced = debounce(saveAll, 300);
+initGcs();
 /* ===== Other UI ===== */
-$('#btnGCS15').addEventListener('click',()=>{
-  $('#d_gksa').value=4; $('#d_gksk').value=5; $('#d_gksm').value=6;
-  ['#d_gksa','#d_gksk','#d_gksm'].forEach(sel=>$(sel).dispatchEvent(new Event('input')));
-  saveAll();
-});
-function setupGcsCalc(prefix){
-  const panel=$(`#${prefix}_gcs_calc`);
-  const selA=$(`#${prefix}_gcs_calc_a`);
-  const selK=$(`#${prefix}_gcs_calc_k`);
-  const selM=$(`#${prefix}_gcs_calc_m`);
-  const apply=$(`#${prefix}_gcs_apply`);
-  const total=$(`#${prefix}_gcs_calc_total`);
-  const btnClose=$(`#${prefix}_gcs_close`);
-  const btn = prefix==='spr' ? $('#btnSprGCSCalc') : $('#btnGCSCalc');
-  if(!panel||!selA||!selK||!selM||!apply||!total) return ()=>{};
-
-  const update=()=>{ total.textContent=gksSum(selA.value,selK.value,selM.value); };
-  [selA,selK,selM].forEach(sel=>sel.addEventListener('change',update));
-
-  let firstFocusable, lastFocusable;
-  const close=()=>{
-    panel.style.display='none';
-    document.removeEventListener('keydown',onKey);
-    document.removeEventListener('click',onDocClick);
-    if(btn) btn.focus();
-  };
-  const onKey=e=>{
-    if(e.key==='Escape') return close();
-    if(e.key==='Tab'){
-      if(e.shiftKey && document.activeElement===firstFocusable){
-        e.preventDefault();
-        lastFocusable.focus();
-      }else if(!e.shiftKey && document.activeElement===lastFocusable){
-        e.preventDefault();
-        firstFocusable.focus();
-      }
-    }
-  };
-  const onDocClick=e=>{ if(!panel.contains(e.target) && e.target!==btn) close(); };
-
-  apply.addEventListener('click',()=>{
-    if(selA.value) $(`#${prefix}_gksa`).value=selA.value;
-    if(selK.value) $(`#${prefix}_gksk`).value=selK.value;
-    if(selM.value) $(`#${prefix}_gksm`).value=selM.value;
-    ['#'+prefix+'_gksa','#'+prefix+'_gksk','#'+prefix+'_gksm'].forEach(sel=>$(sel).dispatchEvent(new Event('input')));
-    close();
-    saveAll();
-  });
-
-  if(btnClose) btnClose.addEventListener('click',close);
-
-  return ()=>{
-    const hidden=panel.style.display==='none';
-    if(hidden){
-      panel.style.display='block';
-      const focusables=panel.querySelectorAll('button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])');
-      firstFocusable=focusables[0];
-      lastFocusable=focusables[focusables.length-1];
-      document.addEventListener('keydown',onKey);
-      document.addEventListener('click',onDocClick);
-      selA.focus();
-    }else{
-      close();
-    }
-  };
-}
-if($('#d_gcs_calc') && $('#btnGCSCalc')){
-  const toggleDGcs=setupGcsCalc('d');
-  $('#btnGCSCalc').addEventListener('click',toggleDGcs);
-}
-if($('#spr_gcs_calc') && $('#btnSprGCSCalc')){
-  const toggleSprGcs=setupGcsCalc('spr');
-  $('#btnSprGCSCalc').addEventListener('click',toggleSprGcs);
-}
 
 function clampNumberInputs(){
   const clamp=el=>{
@@ -237,18 +139,21 @@ function clampNumberInputs(){
 async function init(){
   initTheme();
   await initTopbar();
-    setupHeaderActions({ validateForm, saveAll });
+  setupHeaderActions({ validateForm, saveAll });
   connectSocket();
   await fetchUsers();
   await initSessions();
-  initTabs();
-  initCollapsibles();
+  if(document.getElementById('tabs')){
+    initTabs();
+    initCollapsibles();
+  }
   bodyMap.init(saveAllDebounced);
   initChips(saveAllDebounced);
   initAutoActivate(saveAllDebounced);
   initActions(saveAllDebounced);
   initTimeline();
   setupActivationControls();
+  initMechanismList();
   document.addEventListener('input', saveAllDebounced);
 
   const vitals = {
@@ -282,42 +187,17 @@ async function init(){
       if(chip && isChipActive(chip)) logEvent('vital', label, chip.dataset.value);
     });
   });
-  const updateCirculationMetrics=()=>{
-    const hr=parseFloat($('#c_hr')?.value);
-    const sbp=parseFloat($('#c_sbp')?.value);
-    const dbp=parseFloat($('#c_dbp')?.value);
-    const mapEl=$('#c_map');
-    const siEl=$('#c_si');
-    const map=!isNaN(sbp)&&!isNaN(dbp)?Math.round((sbp+2*dbp)/3):'';
-    const si=!isNaN(hr)&&!isNaN(sbp)&&sbp>0?(hr/sbp).toFixed(2):'';
-    if(mapEl) mapEl.textContent=map;
-    if(siEl) siEl.textContent=si;
-  };
-  initCirculationChecks(updateCirculationMetrics);
-  updateCirculationMetrics();
-    const updateDGksTotal=()=>{
-      $('#d_gks_total').textContent=gksSum($('#d_gksa').value,$('#d_gksk').value,$('#d_gksm').value);
-    };
-    ['#d_gksa','#d_gksk','#d_gksm'].forEach(sel=>$(sel).addEventListener('input', updateDGksTotal));
-    const updateGmpGksTotal=()=>{
-      $('#gmp_gks_total').textContent=gksSum($('#gmp_gksa').value,$('#gmp_gksk').value,$('#gmp_gksm').value);
-    };
-    ['#gmp_gksa','#gmp_gksk','#gmp_gksm'].forEach(sel=>$(sel).addEventListener('input', updateGmpGksTotal));
-    if($('#spr_gks_total')){
-      const updateSprGksTotal=()=>{
-        $('#spr_gks_total').textContent=gksSum($('#spr_gksa').value,$('#spr_gksk').value,$('#spr_gksm').value);
-      };
-      ['#spr_gksa','#spr_gksk','#spr_gksm'].forEach(sel=>$(sel).addEventListener('input', updateSprGksTotal));
-    }
-    $('#btnGmpNow').addEventListener('click', ()=>{ $('#gmp_time').value=nowHM(); saveAll(); });
-    $('#btnSprNow').addEventListener('click', ()=>{ $('#spr_time').value=nowHM(); saveAll(); });
-  $('#btnOxygen').addEventListener('click', ()=>{
-    const box = $('#oxygenFields');
+
+  initCirculation();
+  const btnOxygen=$('#btnOxygen');
+  if(btnOxygen) btnOxygen.addEventListener('click',()=>{
+    const box=$('#oxygenFields');
     box.classList.toggle('hidden');
     saveAll();
   });
-  $('#btnDPV').addEventListener('click', ()=>{
-    const box = $('#dpvFields');
+  const btnDPV=$('#btnDPV');
+  if(btnDPV) btnDPV.addEventListener('click',()=>{
+    const box=$('#dpvFields');
     box.classList.toggle('hidden');
     saveAll();
   });
@@ -329,16 +209,20 @@ async function init(){
     });
     $('#output').addEventListener('input', expandOutput);
     loadAll();
-    ensureSingleTeam();
-    updateActivationIndicator();
+    if(document.getElementById('chips_red') && document.getElementById('chips_yellow')){
+      ensureSingleTeam();
+      updateActivationIndicator();
+    }
     expandOutput();
     clampNumberInputs();
     initValidation();
     validateVitals();
-    updateDGksTotal();
-    updateGmpGksTotal();
   }
-  window.addEventListener('DOMContentLoaded', init);
+  if(document.readyState==='loading'){
+    window.addEventListener('DOMContentLoaded', init);
+  }else{
+    init();
+  }
 
 function validateForm(){
   const fields=[
