+++ docs/js/config.js
@@ -1,0 +1,12 @@
+export const IMG_CT = ['Galvos KT','Kaklo KT','Viso kūno KT'];
+export const IMG_XRAY = ['Krūtinės Ro','Dubens Ro'];
+export const LABS = ['BKT','Biocheminis tyrimas','Krešumai','Fibrinogenas','ROTEM','Kraujo grupė','Kraujo dujos'];
+export const BLOOD_GROUPS = ['0-','0+','A-','A+','B-','B+','AB-','AB+'];
+export const FAST_AREAS = [
+  {name:'Perikardas', marker:'skystis'},
+  {name:'Dešinė pleura', marker:'skystis ar oras'},
+  {name:'Kairė pleura', marker:'skystis ar oras'},
+  {name:'RUQ', marker:'skystis'},
+  {name:'LUQ', marker:'skystis'},
+  {name:'Dubuo', marker:'skystis'},
+];
