+++ public/js/app.js
@@ -1,4 +1,4 @@
-import { $, $$, nowHM } from './utils.js';
+import { $, $$ } from './utils.js';
 import { initTabs } from './tabs.js';
 import { initChips, setChipActive, isChipActive, addChipIndicators } from './chips.js';
 import { initAutoActivate } from './autoActivate.js';
@@ -11,17 +11,16 @@ import { initTopbar } from './components/topbar.js';
 import { initCollapsibles } from './sections.js';
 import { connectSocket, initSessions, fetchUsers, initTheme, saveAll, loadAll } from './sessionManager.js';
 import bodyMap from './bodyMap.js';
-import { generateReport, gksSum } from './report.js';
+import { generateReport } from './report.js';
 import { setupHeaderActions } from './headerActions.js';
 import { TEAM_ROLES } from './constants.js';
-import { initCirculationChecks } from './circulation.js';
+import { initCirculation } from './circulation.js';
+import { setupActivationControls, ensureSingleTeam, updateActivationIndicator } from './activation.js';
+import { initGcs } from './gcs.js';
+import { IMG_CT, IMG_XRAY, LABS, BLOOD_GROUPS, FAST_AREAS } from './config.js';
 export { validateVitals };
 
 /* ===== Imaging / Labs / Team ===== */
-const IMG_CT=['Galvos KT','Kaklo KT','Viso kūno KT'];
-const IMG_XRAY=['Krūtinės Ro','Dubens Ro'];
-const LABS=['BKT','Biocheminis tyrimas','Krešumai','Fibrinogenas','ROTEM','Kraujo grupė','Kraujo dujos'];
-const BLOOD_GROUPS=['0-','0+','A-','A+','B-','B+','AB-','AB+'];
 const LS_MECHANISM_KEY='traumos_mechanizmai';
 
 const imgCtWrap=$('#imaging_ct'); IMG_CT.forEach(n=>{const s=document.createElement('span'); s.className='chip'; s.dataset.value=n; s.textContent=n; addChipIndicators(s); imgCtWrap.appendChild(s);});
@@ -53,16 +52,8 @@ if(bloodUnitsInput && bloodGroupWrap && addBloodOrderBtn){
   addBloodOrderBtn.addEventListener('click',addOrder);
   bloodUnitsInput.addEventListener('keydown',e=>{ if(e.key==='Enter') addOrder(e); });
 }
-const fastAreas=[
-  {name:'Perikardas', marker:'skystis'},
-  {name:'Dešinė pleura', marker:'skystis ar oras'},
-  {name:'Kairė pleura', marker:'skystis ar oras'},
-  {name:'RUQ', marker:'skystis'},
-  {name:'LUQ', marker:'skystis'},
-  {name:'Dubuo', marker:'skystis'}
-];
 const fastWrap=$('#fastGrid');
-fastAreas.forEach(({name,marker})=>{
+FAST_AREAS.forEach(({name,marker})=>{
   const box=document.createElement('div');
   box.innerHTML=`<label>${name} (${marker})</label><div class="row"><label class="pill red"><input type="radio" name="fast_${name}" value="Yra"> Yra</label><label class="pill"><input type="radio" name="fast_${name}" value="Nėra"> Nėra</label></div>`;
   fastWrap.appendChild(box);
@@ -100,62 +91,6 @@ function initMechanismList(){
   });
 }
 
-/* ===== Activation indicator ===== */
-function ensureSingleTeam(){
-  const red=$('#chips_red');
-  const yellow=$('#chips_yellow');
-  const redActive=$$('.chip.active', red).length>0;
-  const yellowActive=$$('.chip.active', yellow).length>0;
-  if(redActive && yellowActive){
-    $$('.chip', yellow).forEach(c=>setChipActive(c,false));
-  }
-}
-function updateActivationIndicator(){
-  const dot=$('#activationIndicator');
-  const status=$('#activationStatusText');
-  const redActive=$$('.chip.active', $('#chips_red')).length>0;
-  const yellowActive=$$('.chip.active', $('#chips_yellow')).length>0;
-  dot.classList.remove('red','yellow');
-  dot.style.display='none';
-  let text='No team activated';
-  if(redActive){
-    dot.classList.add('red');
-    dot.style.display='inline-block';
-    text='Red team activated';
-  } else if(yellowActive){
-    dot.classList.add('yellow');
-    dot.style.display='inline-block';
-    text='Yellow team activated';
-  }
-  if(status) status.textContent=text;
-}
-function setupActivationControls(){
-  const redGroup=$('#chips_red');
-  const yellowGroup=$('#chips_yellow');
-  redGroup.addEventListener('click',e=>{
-    const chip=e.target.closest('.chip');
-    if(!chip) return;
-    if(isChipActive(chip)){
-      $$('.chip', yellowGroup).forEach(c=>setChipActive(c,false));
-    }
-    ensureSingleTeam();
-    updateActivationIndicator();
-    saveAll();
-  });
-  yellowGroup.addEventListener('click',e=>{
-    const chip=e.target.closest('.chip');
-    if(!chip) return;
-    if(isChipActive(chip)){
-      $$('.chip', redGroup).forEach(c=>setChipActive(c,false));
-    }
-    ensureSingleTeam();
-    updateActivationIndicator();
-    saveAll();
-  });
-}
-window.updateActivationIndicator=updateActivationIndicator;
-window.ensureSingleTeam=ensureSingleTeam;
-
 /* ===== Save / Load ===== */
 function expandOutput(){
   const ta = $('#output');
@@ -172,81 +107,8 @@ function debounce(fn, delay){
   };
 }
 const saveAllDebounced = debounce(saveAll, 300);
+initGcs();
 /* ===== Other UI ===== */
-$('#btnGCS15').addEventListener('click',()=>{
-  $('#d_gksa').value=4; $('#d_gksk').value=5; $('#d_gksm').value=6;
-  ['#d_gksa','#d_gksk','#d_gksm'].forEach(sel=>$(sel).dispatchEvent(new Event('input')));
-  saveAll();
-});
-function setupGcsCalc(prefix){
-  const panel=$(`#${prefix}_gcs_calc`);
-  const selA=$(`#${prefix}_gcs_calc_a`);
-  const selK=$(`#${prefix}_gcs_calc_k`);
-  const selM=$(`#${prefix}_gcs_calc_m`);
-  const apply=$(`#${prefix}_gcs_apply`);
-  const total=$(`#${prefix}_gcs_calc_total`);
-  const btnClose=$(`#${prefix}_gcs_close`);
-  const btn = prefix==='spr' ? $('#btnSprGCSCalc') : $('#btnGCSCalc');
-  if(!panel||!selA||!selK||!selM||!apply||!total) return ()=>{};
-
-  const update=()=>{ total.textContent=gksSum(selA.value,selK.value,selM.value); };
-  [selA,selK,selM].forEach(sel=>sel.addEventListener('change',update));
-
-  let firstFocusable, lastFocusable;
-  const close=()=>{
-    panel.style.display='none';
-    document.removeEventListener('keydown',onKey);
-    document.removeEventListener('click',onDocClick);
-    if(btn) btn.focus();
-  };
-  const onKey=e=>{
-    if(e.key==='Escape') return close();
-    if(e.key==='Tab'){
-      if(e.shiftKey && document.activeElement===firstFocusable){
-        e.preventDefault();
-        lastFocusable.focus();
-      }else if(!e.shiftKey && document.activeElement===lastFocusable){
-        e.preventDefault();
-        firstFocusable.focus();
-      }
-    }
-  };
-  const onDocClick=e=>{ if(!panel.contains(e.target) && e.target!==btn) close(); };
-
-  apply.addEventListener('click',()=>{
-    if(selA.value) $(`#${prefix}_gksa`).value=selA.value;
-    if(selK.value) $(`#${prefix}_gksk`).value=selK.value;
-    if(selM.value) $(`#${prefix}_gksm`).value=selM.value;
-    ['#'+prefix+'_gksa','#'+prefix+'_gksk','#'+prefix+'_gksm'].forEach(sel=>$(sel).dispatchEvent(new Event('input')));
-    close();
-    saveAll();
-  });
-
-  if(btnClose) btnClose.addEventListener('click',close);
-
-  return ()=>{
-    const hidden=panel.style.display==='none';
-    if(hidden){
-      panel.style.display='block';
-      const focusables=panel.querySelectorAll('button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])');
-      firstFocusable=focusables[0];
-      lastFocusable=focusables[focusables.length-1];
-      document.addEventListener('keydown',onKey);
-      document.addEventListener('click',onDocClick);
-      selA.focus();
-    }else{
-      close();
-    }
-  };
-}
-if($('#d_gcs_calc') && $('#btnGCSCalc')){
-  const toggleDGcs=setupGcsCalc('d');
-  $('#btnGCSCalc').addEventListener('click',toggleDGcs);
-}
-if($('#spr_gcs_calc') && $('#btnSprGCSCalc')){
-  const toggleSprGcs=setupGcsCalc('spr');
-  $('#btnSprGCSCalc').addEventListener('click',toggleSprGcs);
-}
 
 function clampNumberInputs(){
   const clamp=el=>{
@@ -277,12 +139,14 @@ function clampNumberInputs(){
 async function init(){
   initTheme();
   await initTopbar();
-    setupHeaderActions({ validateForm, saveAll });
+  setupHeaderActions({ validateForm, saveAll });
   connectSocket();
   await fetchUsers();
   await initSessions();
-  initTabs();
-  initCollapsibles();
+  if(document.getElementById('tabs')){
+    initTabs();
+    initCollapsibles();
+  }
   bodyMap.init(saveAllDebounced);
   initChips(saveAllDebounced);
   initAutoActivate(saveAllDebounced);
@@ -324,37 +188,7 @@ async function init(){
     });
   });
 
-  const updateCirculationMetrics=()=>{
-    const hr=parseFloat($('#c_hr')?.value);
-    const sbp=parseFloat($('#c_sbp')?.value);
-    const dbp=parseFloat($('#c_dbp')?.value);
-    const mapEl=$('#c_map');
-    const siEl=$('#c_si');
-    const map = !isNaN(sbp) && !isNaN(dbp) ? Math.round((sbp + 2*dbp)/3) : '';
-    const si = !isNaN(hr) && !isNaN(sbp) && sbp>0 ? (hr/sbp).toFixed(2) : '';
-    if(mapEl) mapEl.textContent=map;
-    if(siEl) siEl.textContent=si;
-  };
-  initCirculationChecks(updateCirculationMetrics);
-  updateCirculationMetrics();
-    const updateDGksTotal=()=>{
-      $('#d_gks_total').textContent=gksSum($('#d_gksa').value,$('#d_gksk').value,$('#d_gksm').value);
-    };
-    ['#d_gksa','#d_gksk','#d_gksm'].forEach(sel=>$(sel).addEventListener('input', updateDGksTotal));
-    const updateGmpGksTotal=()=>{
-      $('#gmp_gks_total').textContent=gksSum($('#gmp_gksa').value,$('#gmp_gksk').value,$('#gmp_gksm').value);
-    };
-    ['#gmp_gksa','#gmp_gksk','#gmp_gksm'].forEach(sel=>$(sel).addEventListener('input', updateGmpGksTotal));
-    if($('#spr_gks_total')){
-      const updateSprGksTotal=()=>{
-        $('#spr_gks_total').textContent=gksSum($('#spr_gksa').value,$('#spr_gksk').value,$('#spr_gksm').value);
-      };
-      ['#spr_gksa','#spr_gksk','#spr_gksm'].forEach(sel=>$(sel).addEventListener('input', updateSprGksTotal));
-    }
-    const btnGmpNow=$('#btnGmpNow');
-    if(btnGmpNow) btnGmpNow.addEventListener('click',()=>{ $('#gmp_time').value=nowHM(); saveAll(); });
-    const btnSprNow=$('#btnSprNow');
-    if(btnSprNow) btnSprNow.addEventListener('click',()=>{ $('#spr_time').value=nowHM(); saveAll(); });
+  initCirculation();
   const btnOxygen=$('#btnOxygen');
   if(btnOxygen) btnOxygen.addEventListener('click',()=>{
     const box=$('#oxygenFields');
@@ -375,16 +209,20 @@ async function init(){
     });
     $('#output').addEventListener('input', expandOutput);
     loadAll();
-    ensureSingleTeam();
-    updateActivationIndicator();
+    if(document.getElementById('chips_red') && document.getElementById('chips_yellow')){
+      ensureSingleTeam();
+      updateActivationIndicator();
+    }
     expandOutput();
     clampNumberInputs();
     initValidation();
     validateVitals();
-    updateDGksTotal();
-    updateGmpGksTotal();
   }
-  window.addEventListener('DOMContentLoaded', init);
+  if(document.readyState==='loading'){
+    window.addEventListener('DOMContentLoaded', init);
+  }else{
+    init();
+  }
 
 function validateForm(){
   const fields=[
